Task 0

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -13,7 +13,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 	if (!table)
 		return (NULL);
 	table->size = size;
-	table->array = calloc((size_t)size, sizeof(hash_node_t *));
+	table->array = calloc(size, sizeof(hash_node_t *));
 	if (!table->array)
 		return (NULL);
 	return (table);
